organize mocking utils and fix issue in certificate reducer
