+++ packages/webcert/src/store/certificate/certificateReducer.ts
@@ -105,7 +105,10 @@ const certificateReducer = createReducer(getInitialState(), (builder) =>
       state.certificateEvents.splice(0, state.certificateEvents.length)
       for (const questionId in state.certificate.data) {
         const question = state.certificate.data[questionId]
-        question.visible = question.visible === undefined ? true : question.visible
+        if (question.visible === undefined) {
+          question.visible = true
+        }
+
         if (question.config.type === ConfigTypes.CATEGORY) {
           continue
         }
