+++ packages/webcert/src/utils/faker/fakeResourceLink.ts
@@ -1,0 +1,13 @@
+import { faker } from '@faker-js/faker'
+import { ResourceLink, ResourceLinkType } from '@frontend/common'
+
+export const fakeResourceLink = (data?: Partial<ResourceLink>): ResourceLink => {
+  const type = data?.type ?? ResourceLinkType.EDIT_CERTIFICATE
+  return {
+    type,
+    name: faker.lorem.word(),
+    description: `${type} - ${faker.lorem.sentence()}`,
+    enabled: true,
+    ...data,
+  }
+}
